Renamed renderer.user.js
